+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## 1.0.0 - In-Progress
 * New plugin module created.
+* Support selecting agreements